Reworked testing and add docker test to package.json
