+++ test/version020/index.js
@@ -1,0 +1,40 @@
+'use strict'
+
+const mmq       = require('../../index.js')
+const chai      = require('chai')
+const should    = chai.should()
+
+describe('Complete test mysql-mq', () => {
+  const store = {
+    queue: null,
+  }
+
+  describe('Prepare', () => {
+    it('Create a Queue (Deprecated)', (done) => {
+      store.queue = new mmq('mmq_test', { database: 'mp_test' })
+      store.queue.init((err) => {
+        should.not.exist(err)
+        done(err)
+      })
+    }).timeout(15000) // High timeout due Travis-CI tests
+  })
+
+  describe('Basic usage', () => {
+    require('./units/basic_004')(store)
+    require('./units/basic_promise')(store)
+    require('./units/queue_operations')(store)
+  })
+
+  describe('Unprepare', () => {
+    it('Close Connection', (done) => {
+      store.queue.end((err) => {
+        should.not.exist(err)
+        done(err)
+      })
+    })
+  })
+
+  describe('Examples', () => {
+    require('./units/examples')
+  })
+})
